V1 Prototype Documentation  (#72)
This PR is meant to cleanup documentation related to the V1 prototype in order to close out the [V1 Milestone Protoype](https://github.com/pokt-network/pocket/milestone/1). It fixes a few minor bugs, but also refactors and updates the documentation so it is easier to follow.

**Fixes issues**
* https://github.com/pokt-network/pocket/issues/21
* https://github.com/pokt-network/pocket/issues/20

**Other fixes**
This change also fixes some issues (protobuf, configuration, etc) that were introduced in the last stages of the integration phase. Anyone cloning the repository anew should be able to run a localnet by following the instructions.

Please mark the options that are relevant.

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] This change requires a documentation update

This change was tested by following the new instructions added to `docs/development/README.md`.

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation
- [ ] No new warnings are generated by my changes
- [X] I have added tests that prove my fix is effective or that my feature works
- [X] New and existing unit tests pass locally with my changes
- [X] Any dependent changes have been merged and published in downstream modules