[Documentation] Remove text-heavy contribution guide (#170)
## Description

Remove the text-heavy contribution guide.

## Type of change

The original contribution guide was built for a live protocol with a lot of changes (e.g. integration branches, milestones, etc...). For a protocol that is not live, this is overhead that may "scare" new contributors from reading a lot of text.

The current contribution guide can always be accessed via an old commit [here](https://github.com/pokt-network/pocket/blob/18581335c954300915a668a6787553aa06b48344/docs/contributing/contribution/development.md) and may be re-introduced once the codebase is more mature. For now, it is easier not to have it be in the codebase.

The contribution

- [ ] New feature (non-breaking change which adds functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] Documentation
- [ ] Other (<_REPLACE_ME_WITH_DETAILS_>)