[Tooling] CHANGELOG and Version Bump PreCommit Hook (#397)
<!--
 1. Make the title of the PR is descriptive and follows this format: `[<Module>] <DESCRIPTION>`
 2. Update the _Assigness_, _Labels_, _Projects_, _Milestone_ before submitting the PR for review.
 3. Add label(s) for the purpose (e.g. `persistence`) and, if applicable, priority (e.g. `low`) labels as well.
-->

## Description

<!--
1. Add a summary of the change including: motivation, reasons, context, dependencies, etc...
2. If applicable, specify the key files that should be looked at.
-->

This PR contains the pre-commit logic to check and validate whether the relevant `CHANGELOG.md` files have (1) been included in the commit and (2) contain up-to-date versioning. This is in response to [Issue #93](https://github.com/pokt-network/pocket/issues/93).

Git Hook:
The client-side logic is contained in the bash script `.githooks/pre-commit` for it to actually perform this verification check prior to a commit it must be copied into `.git/hooks` and be executable.
```sh
cp .githooks/pre-commit .git/hooks
chmod +x .git/hooks/pre-commit
```
As it is a git hook, it can be disabled locally by running `git commit` with the `--no-verify` or `-n` flag:
```sh
git commit -n ...
```
The script will detect if a `CHANGELOG.md` is present and if it is it will check whether (1) the latest version in the changelog is greater than the one before it and (2) that the date alongside the latest version is the current date. If any of these checks fail (no changelog, invalid version, not current date) then the script will block the commit.

To verify the changelog's version is correct it converts the string `[0.0.0.8]` to `0008` and then strips the leading zeros and compares the number against the next version found if it is less than or equal to the one previous it sees this as an error. This is done by extracting the versions and dates from the `CHANGELOG.md` file using `sed`. It then compares the date of the newest version against today's date and if they aren't equal and it sees this as an error.

Currently, the script uses the directories `.githooks` `.github` `docs` `bin` as ignored directories and does not check for a `CHANGELOG.md` file in these subdirs.

Dependencies:
This script requires `bash` to be installed and requires `sed`, `head` and `date` commands - these should be present on all Linux/OSX/BSD systems. The script uses sed Basic Regular Expressions that are portable across GNU sed and FreeBSD sed (used by OSX as well) and also uses `date` formatting that is consistent across Linux/OSX/BSD versions of the command. Currently, there is no support for Windows when run locally.


GitHub Workflow:
The GitHub workflow will run the `pre-commit` script however it will populate the edited files by instead passing them as arguments to the script. This cannot be disabled and will run on all PRs when they are opened, reopened and synchronised. The jobs are configured to run on `ubuntu-latest` so there are no issues with dependencies as all the required binaries are present on ubuntu installs.

The files in question are found in:
```
.github
└── [ 112] workflows
      └── [ 1KB] changelog-verify.yml
.githooks
└── [ 4KB] pre-commit
docs
└── [ 60] development
      └── [ 6KB] README.md
```


## Issue

Fixes #93 

## Type of change

Please mark the relevant option(s):

- [x] New feature, functionality or library
- [ ] Bug fix
- [ ] Code health or cleanup
- [ ] Major breaking change
- [ ] Documentation
- [ ] Other <!-- add details here if it a different type of change -->

## List of changes

<!-- List out all the changes made-->

- Add `pre-commit` git hook to validate changelogs presence in commits & correct versioning
- Add GitHub workflow to run the `pre-commit` logic on all PRs [opened, reopened, synchronised]
- Add a guide to activate `pre-commit` hook to `docs/development/README.md`
- Updated `utility/doc/CHANGELOG.md`

## Testing

<!--
If you added additional tests or infrastructure, describe it here.

Bonus points for images and videos or gifs.
-->
Videos:
[changelog-verify-test-video.webm](https://user-images.githubusercontent.com/53987565/207724785-71f2bcc1-f3a5-401e-b543-17efaadaa6f2.webm)
[github-workflow-test-video1.webm](https://user-images.githubusercontent.com/53987565/207724787-b7c7c303-ab22-4689-a13c-341ef1cd7901.webm)
[github-workflow-test-video2.webm](https://user-images.githubusercontent.com/53987565/207724793-ed4c2b38-5c57-407d-97c0-98999ba11d4f.webm)


## Required Checklist

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have tested my changes using the available tooling --> I have tested locally, and using GitHub PRs for the workflow
- [x] I have updated the corresponding CHANGELOG

### If Applicable Checklist
- [x] I have updated the corresponding README(s); local and/or global
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added, or updated, [mermaid.js](https://mermaid-js.github.io) diagrams in the corresponding README(s)
- [ ] I have added, or updated, documentation and [mermaid.js](https://mermaid-js.github.io) diagrams in `shared/docs/*` if I updated `shared/*`README(s)