[Tooling] fix: changelog verification workflow (#468)
## Description

While recent update to changelog verification works as expected when run
locally, it seems to consistently return a false positive result when
running in GitHub actions. Testing @h5law's
[hypothesis](https://discordapp.com/channels/553741558869131266/986789914379186226/1067940783111544952)
(i.e. maybe `HEAD` or `main` aren't what we expect them to be in that
environment). TLDR; we probably shouldn't rely on git refs in this
context.

This PR addresses fixes this false positive scenario as well as cleans
up a bit.

## Issue

No issue as of time of writing, see:
https://discordapp.com/channels/553741558869131266/986789914379186226/1067940783111544952

## Type of change

Please mark the relevant option(s):

- [ ] New feature, functionality or library
- [x] Bug fix
- [x] Code health or cleanup
- [ ] Major breaking change
- [ ] Documentation
- [ ] Other <!-- add details here if it a different type of change -->

## List of changes

- Rename pre-push to pre-receive: `pre-push` conventionally receives 2
arguments (remote and url), which aren't useful for our case;
additionally, `pre-receive` is a better conceptual fit for the purpose
it serves
- Receive list of changed files as argument instead of trying to compute
in the same script
- Add quotes around changed files variable interpolation

## Testing

- [ ] `make develop_test`
- [ ]
[LocalNet](https://github.com/pokt-network/pocket/blob/main/docs/development/README.md)
w/ all of the steps outlined in the `README`
- [x] Tested via [PR to my
fork](https://github.com/bryanchriswhite/pocket/pull/3)


## Required Checklist

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have tested my changes using the available tooling
- [ ] I have updated the corresponding CHANGELOG

### If Applicable Checklist
- [ ] I have updated the corresponding README(s); local and/or global
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added, or updated,
[mermaid.js](https://mermaid-js.github.io) diagrams in the corresponding
README(s)
- [ ] I have added, or updated, documentation and
[mermaid.js](https://mermaid-js.github.io) diagrams in `shared/docs/*`
if I updated `shared/*`README(s)

---------

Co-authored-by: harry <hrryslw@pm.me>