[Persistence] Add deterministic `Hash` functions to `Param` and `Flag` protobufs (#472)
## Description

This PR is in response to and issue where the local `make test_all`
command will pass and the GitHub worklow test will also pass, but at
times will fail on the
`TestStateHash_DeterministicStateWhenUpdatingAppStake` test.


![image](https://user-images.githubusercontent.com/53987565/215277664-e38de2fa-880d-4d62-8bae-02b54a7c233e.png)

The `String()` function generated by a protobuf to output its contents
in a `string` format is not deterministic. As such the fields can appear
in different orders. This would mean that the resulting hashing of these
`string` values can differ if run numerous times.

When this function is used to update the `param` and `flags` trees their
is a possibility that the hashes will be different to those hard-coded
in the test case.

To fix this the marshalled protobuf's bytes are now being hashed using
`crypto.SHA3Hash()` and used as the keys in the tree when updating. As
these marshalled protobufs are using the `deterministic: true` option
this will result in a deterministic hashing for the keys being used.


## Issue

Fixes N/A

## Type of change

Please mark the relevant option(s):

- [ ] New feature, functionality or library
- [x] Bug fix
- [ ] Code health or cleanup
- [ ] Major breaking change
- [ ] Documentation
- [ ] Other <!-- add details here if it a different type of change -->

## List of changes

- Use `crypto.SHA3Hash(protobufBz)` for the keys in the `Param` and
`Flag` tree update methods
- Update hard-coded hashes in the
`TestStateHash_DeterministicStateWhenUpdatingAppStake` test
- Use lowercase names for fields in the `Param` and `Flag` message
definitions to fit convention

## Testing

- [x] `make develop_test`
- [x]
[LocalNet](https://github.com/pokt-network/pocket/blob/main/docs/development/README.md)
w/ all of the steps outlined in the `README`

## Required Checklist

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have tested my changes using the available tooling
- [x] I have updated the corresponding CHANGELOG

### If Applicable Checklist
- [ ] I have updated the corresponding README(s); local and/or global
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added, or updated,
[mermaid.js](https://mermaid-js.github.io) diagrams in the corresponding
README(s)
- [ ] I have added, or updated, documentation and
[mermaid.js](https://mermaid-js.github.io) diagrams in `shared/docs/*`
if I updated `shared/*`README(s)