[Client Tooling] Keybase (issue #455) (#459)
## Description

This is an implementation of a Keybase for the pocket V1 client.

It uses `BadgerDB` as the local database for the persistent storage of
keys. The current `Keybase` interface has the following methods


![image](https://user-images.githubusercontent.com/53987565/214436549-efcf0895-ddf1-477a-8624-225ab094c713.png)

These are covered with the following tests


![image](https://user-images.githubusercontent.com/53987565/214436651-c32da614-df53-416d-9753-25c8fcb261eb.png)

The tests use an in-memory database created with `NewKeybaseInMemory()`
whereas when exposed to the user the CLI should call `NewKeybase()`
instead with a default path/user-supplied path. The CLI integration has
not been implemented as part of this PR however.

### How it works

The keybase heavily relies upon `shared/crypto/ed25519.go` and the
`PublicKey` and `PrivateKey` interfaces it exposes.

The keys are created using the methods from this library. Keys are then
encrypted and armoured in the same fashion as in V0 and stored in the DB
as a `KeyPair` struct (encoded to a `[]byte`).


![image](https://user-images.githubusercontent.com/53987565/214108865-bdb8f184-a1ff-4524-b978-af6e2b8a41c9.png)

This struct contains the public key and the encrypted JSON encoded
private key string. This struct has the following methods:
- `GetAddressBytes() []byte` -> returns `[]byte` of the public key
address
- `GetAddressString() string` -> returns the hex `string` of the public
key address
- `Unarmour(passphrase string) (crypto.PrivateKey, error)` -> returns
the unencrypted unarmoured private key if passphrase is correct
- `ExportString(passphrase string) (string, error)` -> returns raw
private key string is passphrase is correct
- `ExportJSON(passphrase string) (string, error)` -> returns json
armoured private key string if the passphrase is correct

Keys are stored in the database in the following manner:
- The keys for the BadgerDB key-value storage are the `[]byte` value
returned by `KeyPair.GetAddressBytes()` aka the `[]byte` address of the
public key
- The values are the `[]byte` encoded `KeyPair` structs of the key

Key encryption/decryption works with or without a passphrase (when no
passphrase is given `""` must still be passed to the function as the
passphrase argument)

For ease of use the hex `string` returned by
`KeyPair.GetAddressString()` is used to access the keys in storage.

Importing and exporting keys in either JSON string format or the private
key hex string is fully interoperable between V1 and V0

Signing and Verification of messages works - and is covered with a Tx in
the tests. I would need to look into the use cases for this more (for
example multisig) as the current implementation is very rudimentary and
in theory should work for producing a signature on any `[]byte` message
but I would like to look into this more.

## Issue

Fixes #455 

## Type of change

Please mark the relevant option(s):

- [x] New feature, functionality or library
- [ ] Bug fix
- [ ] Code health or cleanup
- [ ] Major breaking change
- [x] Documentation
- [ ] Other <!-- add details here if it a different type of change -->

## List of changes

- Implement a keybase with BadgerDB local DB
- Add functionality to Import/Export keys, Get/List keys, Create new
keys, Delete Keys, Sign Messages, Verify Signatures
- Add `make test_app` entry point in Makefile
- Add unit tests to cover Keybase use cases
- Add documentation for Keybase

## Testing

- [x] `make develop_test`
- [x]
[LocalNet](https://github.com/pokt-network/pocket/blob/main/docs/development/README.md)
w/ all of the steps outlined in the `README`

## Required Checklist

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have tested my changes using the available tooling
- [x] I have updated the corresponding CHANGELOG

### If Applicable Checklist
- [x] I have updated the corresponding README(s); local and/or global
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have added, or updated,
[mermaid.js](https://mermaid-js.github.io) diagrams in the corresponding
README(s)
- [x] I have added, or updated, documentation and
[mermaid.js](https://mermaid-js.github.io) diagrams in `shared/docs/*`
if I updated `shared/*`README(s)

---------

Co-authored-by: Daniel Olshansky <olshansky.daniel@gmail.com>