[Tooling] Integrate Keybase w/ CLI (Issue #484 ) (#501)
## Description

This PR addresses the `#150` comments throughout the `app/client/cli`
directory and integrates the keybase to be used in place of reading keys
from files locally.

The new method has been added to the `debug/keystore.go` file
`InitialiseDebugKeybase()` which reads the 999 validator private keys
from the `build/localnet/manifests/private-keys.yaml` file and stores
them in the keybase. This is only run when there are less than 999 keys
in the keybase. To speed this up `BadgerDB`'s `WriteBatch()` function is
being used - however due to the encryption/JSON encoding and then Gob
encoding steps of adding private keys to the keybase this takes a few
seconds when ran for the first time.

This PR does not add any endpoints to interact with the keybase from the
CLI this is instead tracked in the issue #489

## Issue

Fixes #484 

## Type of change

Please mark the relevant option(s):

- [x] New feature, functionality or library
- [ ] Bug fix
- [x] Code health or cleanup
- [ ] Major breaking change
- [ ] Documentation
- [ ] Other <!-- add details here if it a different type of change -->

## List of changes

- Add logic to initialise keybase with 999 validator keys
- Update `app/client/cli` to use the keybase


## Testing

- [x] `make develop_test`
- [x]
[LocalNet](https://github.com/pokt-network/pocket/blob/main/docs/development/README.md)
w/ all of the steps outlined in the `README`


[localnet_with_integration.webm](https://user-images.githubusercontent.com/53987565/217543043-21c35c3e-8389-44a5-968a-d3afb6cb8dd9.webm)

## Required Checklist

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have tested my changes using the available tooling
- [x] I have updated the corresponding CHANGELOG

### If Applicable Checklist

- [ ] I have updated the corresponding README(s); local and/or global
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added, or updated,
[mermaid.js](https://mermaid-js.github.io) diagrams in the corresponding
README(s)
- [ ] I have added, or updated, documentation and
[mermaid.js](https://mermaid-js.github.io) diagrams in `shared/docs/*`
if I updated `shared/*`README(s)