[Tooling] SLIP-0010 HD Child Key Generation (#510)
## Description

This PR contains an implementation of
[SLIP-0010](https://github.com/satoshilabs/slips/blob/master/slip-0010.md)
which allows for deterministic child key generation from a master key.
This is done by using the seed of the master key and creating a `hmac`
hash using the `sha512` hashing function. The hash is split into 2
`[]byte` slices representing the `SecretKey` and the `ChainCode`.

This `hmac` key can then be used to deterministically generate keys
within the BIP-44 path `m/44'/635'/%d'`. `44` referring to the BIP-44
hardened key purpose of the path and `635'` is the coin type for Pocket
Network. `%d` can be any number between `0-2147483647` inclusive.
Allowing for a large number of child keys to be deterministically
generated and retrieved by index from a single master key's seed.

_NOTE_ All the child keys are ed25519 keys created from the `SecretKey`
of the SLIP HMAC keypair. They are then converted into an encrypted
`KeyPair` interface following the same pattern already found in the
keybase.

Some documentation on this process can be seen
[here](https://github.com/pokt-network/pocket/blob/slip-0010-keys/shared/crypto/README.md#slip-0010-hd-child-key-generation)

The complete test vectors from the SLIP-0010 spec are included in
`slip_test.go` as well as some pokt specific test cases.

The `DeriveChild` CLI endpoint has also been added to allow to end users
to use the SLIP integration.

## Issue

Fixes N/A

## Type of change

Please mark the relevant option(s):

- [x] New feature, functionality or library
- [ ] Bug fix
- [ ] Code health or cleanup
- [ ] Major breaking change
- [x] Documentation
- [ ] Other <!-- add details here if it a different type of change -->

## List of changes

- Implement SLIP-0010
- Integrate child key generation with the keybase
- Add unit tests for child generation and to cover the SLIP-0010 spec
test vectors
- Add CLI endpoint `DeriveChild`
- Update documentation

## Testing

- [x] `make develop_test`
- [x]
[LocalNet](https://github.com/pokt-network/pocket/blob/main/docs/development/README.md)
w/ all of the steps outlined in the `README`

## Required Checklist

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have tested my changes using the available tooling
- [x] I have updated the corresponding CHANGELOG

### If Applicable Checklist

- [x] I have updated the corresponding README(s); local and/or global
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have added, or updated,
[mermaid.js](https://mermaid-js.github.io) diagrams in the corresponding
README(s)
- [ ] I have added, or updated, documentation and
[mermaid.js](https://mermaid-js.github.io) diagrams in `shared/docs/*`
if I updated `shared/*`README(s)