[Utility][RPC][CLI] Querying governance parameters (Issue #619) (#622)
## Description

This PR introduces the ability to query `v1/query/allChainParmas` with a
`GET` request in order to retrieve the current values for all governance
parameters stored in the node's PostgreSQL database. This is defined in
the OpenAPI YAML RPC specification and its logic is implemented in
`rpc/handlers.go`

The CLI endpoint is a subcommand under the `System` command that allows
for the querying of this RPC endpoint and is used by the make target
`query_node_params` to quickly enable the querying of governance
parameters on localnet.

In addition to this the `GovParamMetadatMap` has been moved to
`shared/utils/gov_utils.go` where it is being initialised and used in
`persistance` and `utility`. The `owner` tags have been added to the
`genesis.proto` file to show which parameters own which, this allows for
an easy lookup and reduces the code footprint of the `getParamOwner`
function greatly.

There has also been a lot of code cleanup in `utility` simplifying
governance parameter accessor functions and their tests

## Issue

Fixes #619 
## Type of change

Please mark the relevant option(s):

- [x] New feature, functionality or library
- [ ] Bug fix
- [x] Code health or cleanup
- [ ] Major breaking change
- [ ] Documentation
- [ ] Other <!-- add details here if it a different type of change -->

## List of changes

- Add `GetAllParams()` function to the persistence read interface to
retrieve the current value of each parameter
- Add the `v1/query/allChainParams` RPC endpoint to the RPC
specification, and implement its logic
- Add the `Query` subcommand in the CLI
- Add the `AllChainParams` subcommand to the Query CLI endpoint to query
`v1/query/allChainParams`
- Add `make query_chain_params` target to easily query the parameters on
a localnet
- Move parameter metadata logic into `shared/utils/gov.go`
- Reduce code footprint of gov parameters, their accessor functions and
tests

## Testing

- [x] `make develop_test`
- [x]
[LocalNet](https://github.com/pokt-network/pocket/blob/main/docs/development/README.md)
w/ all of the steps outlined in the `README`

## Required Checklist

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have added, or updated, [`godoc` format
comments](https://go.dev/blog/godoc) on touched members (see:
[tip.golang.org/doc/comment](https://tip.golang.org/doc/comment))
- [x] I have tested my changes using the available tooling
- [x] I have updated the corresponding CHANGELOG

### If Applicable Checklist

- [ ] I have updated the corresponding README(s); local and/or global
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added, or updated,
[mermaid.js](https://mermaid-js.github.io) diagrams in the corresponding
README(s)
- [ ] I have added, or updated, documentation and
[mermaid.js](https://mermaid-js.github.io) diagrams in `shared/docs/*`
if I updated `shared/*`README(s)