[IBC] Add nil check on proof for membership and non-membership proof creation (#877)
## Problem:
In `proofs_ics23.go`, the `createMembershipProof` and `createNonMembershipProof` functions call `tree.Prove(key)` and then pass the `proof` to the `convertSMPToExistenceProof` or `convertSMPToExclusionProof` functions without checking if `proof` is `nil`. This could lead to a panic due to a nil pointer dereference if `tree.Prove(key)` returns a `nil` proof.

## Solution:
This PR adds a check to verify if `proof` is `nil` before passing it to the `convertSMPToExistenceProof` or `convertSMPToExclusionProof` functions. If `proof` is `nil`, an error is returned.

## Changes:
Added nil check on `proof` in `createMembershipProof` and `createNonMembershipProof` functions in `proofs_ics23.go`.